feat: install requirements by default
When running unit or integration tests, install the Ansible dependencies defined in
the corresponding `requirements.yml` file. The semantics of the `-R` parameter has
been inverted, and its long form has been renamed to `--skip-requirements`. As the
suggests, when passed, that option skips the installation of those requirements.

The `vagrant` action is affected - requirements should be installed when running
andebox inside the VM, so the `-R` has been removed from that execution.