feat(action/ignores): remove restriction to spec parameter
Parameter was previously restricted t a list of choices based on the existing filenames.
Now not only that restriction is completely gone, but the parameter itself has been renamed from `--ignore-file-spec` (or `-ifs`) to `--spec` (or `-s`).
Tests have been updated and the two xfail tests are now passing.
Dependency `looseversion` is no lonoger used and has been removed from the project.